AplicacionWebSBPractica2/api-flask/requirements.txt
flask==3.0.2
flask-cors==4.0.0           # para que Spring pueda llamar sin CORS
mysql-connector-python==8.3.0
requests==2.31.0            # para llamar a la PokeAPI